pyproject.toml
[project]
name = "mlbstats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mlb-statsapi>=1.9.0",
    "pydantic>=2.11.3",
    "requests>=2.32.3",
]
